Changed the Anxious Merchant mission to use the same formula as rush cargo.
It was creating frankly ridiculous rewards, likely because there was
a part to the formula I didn't realize I needed to update, but its
formula was rather weird. Better to use the rush cargo formula with
some tweaks (namely: it's always Emergency level, and it awards a
bit more credits).