Added ability to call faction.dynAdd multiple times no problem.
I thought this was how it worked, but apparently not. Decided to
change this, because the alternative is expecting Lua scripters to
keep track of a faction but then know to discard it when changing
systems, which would quickly become a horrible mess. Also documented
how the dynamic faction creation and cleanup works better.