One addition to the OARS rating.
In truth, most people aren't going to consider what I'm actually
referring to as references to slavery, because to most people, "slavery"
is an old thing that we used to do and obviously wrong. Yet, fact
of the matter is prisons *are* institutions of slave labor. Slavery
is *common* all over the world today, and it's not just "third-world
countries" doing it.

Most people are going to misunderstand what I mean when I put "slavery"
as a thing that is in Naikari, but frankly, the fact that many planets
are described as prisons where prisoners are forced to work in mines,
etc *should* be just as shocking as anything with the word "slave" in
it. It's literally exactly the same as how slavery has worked
throughout most of human history. As a society, we're against the
*word* "slave". We think of "slavery" as being that thing "backwards"
countries do, while being absolutely fine with the prison system. It's
patronizing and hypocritical.