Fixed Collective refueling code.
Looks like pilots are initially created while the player doesn't
exist, which makes code that depends on the player rather… fragile.
It's less noticeable with factions other than the Collective, but
maybe something to act on for them as well. Honestly, I'm not so
sure using ai.getstanding() even makes much sense.