Removed all instances of AI init code relying on the player existing.
Most of these cases are easily dealt with by faction.getPlayerStanding(),
and some of them were simply causing AI ships to… not do things for
no reason, when the player didn't exist.

I'll need to do a fly-through of the entire galaxy to make absolutely
sure I didn't break one of them, but this not only seems to work
absolutely fine, it works *better* than before, due to the fact that
the previous method failed for all pilots which spawned during the
simulation period prior to the player pilot being created. Having
seen what I've seen, I truly am baffled that anyone thought the
previous implementation was a good idea. No wonder it was a common
occurrence to have pilots just randomly say, "Sorry, I'm busy now."
It seems that was simply because the initialization wasn't being
done completely, for no good reason.