De-nerfed the Collective drones.
This change probably could have gone into version 0.5, but I didn't
think of it in time. Either way, this makes the drones actually get
equipped with *good* outfits, and not equipped with bad stuff on
purpose.

I was surprised to find this, but it seems the Collective drones are
best off with armor strength instead of cloaking. In hindsight, given
their incredibly close-range fighting, that makes perfect sense.