Made Seek and Destroy "cold" numbers more lenient, no clues from Collective.
The previous numbers for "nbsys" (maximum number of systems before
the trail goes cold) had a high probability of making it so that you
only got one chance, so I upped that so you always have at least
three chances. Also made it so that the Collective never provides
clues; this didn't matter in the past, but you can now hail Collective
drones if you befriend them, so a special case is needed now.