Added special handling for zero-swivel into the aim lines code.
The previous implementation actually would have cause a divide by
zero, but also, it would have led to a less than ideal display even
if it didn't do that. For one, suppose you have two weapons, one
with swivel and one without; the previous implementation would have
taken the lowest common denominator, leading to no visual indicator
of the swivel one of them has, despite the fact that this doesn't
do anything to help with aiming the one that has no swivel. Also,
in the case that a zero-swivel weapon is all you have, it makes no
sense to draw the arc lines at all.

The changes made here take care of all of these issues.