Hugely increased the max delay specified by scom.calcNextSpawn.
I don't think we need to force a spawn every 60 seconds, and in fact
that could be dangerous considering that a lot of spawn targets
(namely carriers) use up a whole lot of presence now. I don't see
a problem with allowing the delay to go really high if necessary to
curb over-spawning.