Fixed some things with ldoc and also functioning of hook.trigger().
It looks like a programming error led to hook.trigger() passing on
one more hook parameter argument than it was supposed to. I believe
the only ill effect of this would have been to offset the argument
passed to the hook by e.g. hook.custom() by one, adding an extra
nil argument in between it and the intended hook parameters. That
was a problem behavior, however.