Removed pilot_cargoRmRaw() from the C API.
pilot_cargoRmRaw() only existed to provide an unused "cleanup"
option which would remove mission cargo; however, not only was there
never a need to do this, the method pilot_cargoRmRaw() used could not
actually do this, as it simply removed the first cargo it saw which
matched the indicated cargo type. This meant, for example, that if
you were carrying regular Ore and mission cargo Ore, trying to remove
Ore would result in one of them being removed at "random" (more
precisely, whichever one was added first would be removed), and this
could lead to dangerous side-effects if pilot_cargoRmRaw() was
depended on for cleanup (as has been the case in Naev for almost a
year until bobbens realized his mistake today).

In practice, pilot_cargoRmRaw() was only ever called by
pilot_cargoRm() with the "ceanup" argument set to false, so I simply
renamed cargoRmRaw() to cargoRm() and removed the "cleanup" argument,
then removed the previous cargoRm() definition and the cargoRmRaw()
declaration in the header.