More or less completely restructured how jump unidiffs are done.
This took literal hours because I got confused by the rather strange
choice to put an essential variable in a union, but anyway. This is
a massive change to the unidiff system stemming from merely a desire
to fix a problem with unidiff-added presences from new jumps not
getting removed when removing that unidiff.

I ended up more or less completely restructuring how the unidiff
system works, and spending hours debugging a problem caused by
attempting to misuse unions. But the end result is that this not only
fixes presence-bleeding not getting reverted properly when a jump
adding diff is removed, it also (at least in theory) should correctly
revert a jump removal diff even if it's on a jump with special flags,
etc.