Changed the jump between C-28 and Mason to a standard, two-way jump.
This causes the Collective to intermingle in Soromid space a bit.
I'm sure bobbens would hate this change, but I think this'll be a
great addition to the mystery of what's going on with the Collective,
since on the Soromid side, they are non-hostile toward everyone
except pirates (in stark contrast with Fortitude).