Changed the snapshot system to use a special extension, not special paths.
Was realizing that it would be nice to be able to have snapshots all
together in the same loading screen, which would have been inordinately
difficult with the approach I established of having sub-directories.
So instead, I've added a new ".ns.snapshot" extension.

This also fixes a couple problems I noticed with the snapshot code.