Use a more robust method to check parent validity in pilot.leader() recursion.
Fixes an uncommon segfault. I don't know what exactly triggers the
segfault, but it's likely that it has something to do with the parent
dying.

This also makes the pilot.leader() function not return a leader that
doesn't exist, as a side-effect, which (judging from my reading of some
of the AI code) is probably a good thing. In fact, I think I'll harden
this even further, so that dead and destroyed pilots also don't count.