Further hardened the checks of pilot.leader() to never return dead pilots.
The effect of this is subtle, but it should slightly improve AI behavior,
due to some places kind of implicitly assuming that a pilot's leader
is always a leader that actually exists.

The whole assignment within conditional idiom isn't something I would
normally use, but that seems to be the best way to keep the code
clean and simple in this case. Too bad this isn't Python code; it'd
be nice to use the walrus operator for the first time!