Only call land_updateMainTab in player.pay() and player.allowSave().
Trying to call land_genWindows() causes some severe problems, among
which I'm aware of it erasing the missions displayed and corrupting the
mission NPCs.

For player.pay() it might be worthwhile to implement an "update all
tabs" function for the land window, but for now, just updating the
main tab (since there's already a public function for that). That's
also all that's needed for the Save Snapshot button, so best to just
keep it nice and simple for now, I think.