Completely removed the now unused afterburner "rumble" feature.
It's not in use and I don't think should be used, given that all it
does is make the game less accessible for a crap visual effect. I've
seen a shake effect for an afterburner that's worked well, but in that
case it was only the GUI that shook, and only subtly. Waving the
screen around messes with your vision a lot more and doesn't look
nearly as good.