A couple small changes to make things a bit better.
I changed the message about not being able to land in es00 to avoid
confusion over what it actually means (which I have seen before), and
I made an edit to how faction_get() works so it's simpler and easier
to read (but without actually changing how it works, at least in
theory).

I was trying to find the source of some warnings that I saw in a
playthrough, but I couldn't figure that out. It doesn't seem to be
easily reproducible.