Started work on some fixes for dynamic factions.
I haven't figured out exactly what causes it yet, but I've noticed
that dynamic factions getting loaded over and over again in a session
seem to cause some sort of weird "ghost factions" to stick around in
some form. This is my first pass at the problem: I made sure that
when dynamic factions are cleared, they also go through their allies
and enemies lists to make sure they get removed as allies and enemies.
This doesn't seem to have solved the problem, but it might be one
piece of the puzzle, at the very least.

At the same time, I made faction name functions in Lua cause an error
if the faction is invalid. The way this is currently happening is not
something it should be doing, but there *are* cases where this error
can be run into by a Lua coder, particularly if they reuse a dynamic
faction variable inappropriately.

I'll need to investigate this some more and see where the rest of
the problems lie.