Changed pilot IDs to use unsigned long, not unsigned int.
unsigned int is only required to be at least 16-bit, supporting at
least values up to 65,535. That's not good enough for this use-case;
if we assume that 100 pilots appear in every system the player enters,
and that the player enters 5 systems a minute, a player could easily
cause this to wrap around (leading to confusing bugs) after only a
couple hours of play. Changing the type to unsigned long guarantees
that the pilot IDs can at least go up to 4,294,967,295, which is
much less likely to be reached.

(In practice, this doesn't change anything. Most platforms will choose
to make the int type 64-bit, or at least 32-bit. But I don't want us
to depend on a non-standard extension of the int type's capabilities.)