Disabled the warnings in faction_isPlayerFriend and _isPlayerEnemy.
After some investigation, I've confirmed that the warnings printed
by these functions were false-positives, so I've decided to simply
disable the warnings, at least for now, and document the reasoning.
Ultimately, it's due to the part of cleanup that calls
pilot_rmHostile(), and while I investigated modifying the code to
prevent the warning, it is ultimately impossible to do so without
creating breakage.

Given the scope of faction_isPlayerFriend() and
faction_isPlayerEnemy(), it's highly unlikely that an invalid faction
would lead to terribly broken behavior (as they both simply return
false in that case, an appropriate response), so disabling the
warning should be perfectly fine. In fact, these functions didn't
have a warning at all before, and areAllies() and areEnemies() were
even warningless.