Removed an unnecessary (and problematic) "cargo space" check in Runaway.
I don't know what the original author of the mission was thinking,
but the only time you would have negative cargo space is if you took
off a cargo pod, or something to that effect, rendering your ship
unspaceworthy. It in any case has nothing to do with adding a 0 mass
mission cargo.