Use cargoAddRaw, not cargoAdd, for mission cargo.
Using cargoAdd() causes the cargo amount to be implicitly changed,
which could mess up missions in unexpected ways. Missions are meant
to be fully in charge of making sure the player has enough cargo
space, so it doesn't make sense to then pass that through something
that could alter it.

This change also means that it's possible for a mission to add more
cargo than the player's ship can carry, rendering the ship not
spaceworthy, which could theoretically be used as a mechanic.