Don't allow pilot_cargoAdd to cause addition of negative cargo.
pilot_cargoAddRaw is not built to accept negative numbers for the
quantity. Doing so would lead to some very strange and buggy behavior.

Undoubtedly, the intention is that the free space is supposed to be
checked before calling pilot_cargoAdd. That said, I don't think it's
helpful for cargoAdd to introduce bugs if code is written that forgets
to do this.