Added another structural slot to the Za'lek Demon.
The loss of a weapon slot seems to be a tremendous detriment compared
to the minor benefit of gaining a utility slot, especially considering
that on a medium ship, the best use of an extra utility slot is just
extra energy for weapons. The added structural slot in addition should
help make the Za'lek Demon more useful.