Changed stats to not show in reverse order, give top billing for main stat.
The previous code caused stats to be listed in reverse order because
it was simpler to program that way, but I've implemented a slightly
more complicated implementation to make it so that the ordering is
identical to what is seen in the file itself. I've also edited outfits
to make sure they list their "primary" stat first in all cases (the one
that serves as their selling point), and have detriments listed below
that.