Fixed inaccurate representation of afterburners' thrust value.
The afterburner's thrust value was reported as a percentage increase,
but this is incorrect. Afterburners in fact have their own independent
thrust value, just as they have their own independent mass limit.
What this effectively does is add a separate acceleration source
which works the same way as normal ship acceleration, just operating
under a different mass limit.

To make this clearer, I've corrected the "Thrust" display to show
acceleration in mAU/s², and without coloring (as this is something
intrinsic to the afterburner, not something that affects the rest
of your ship or its stats).

Maximum speed, on the other hand, is in fact a percentage of your
ship's normal maximum speed.