Made Seek and Destroy update the current system immediately.
It was updating the current system in a delayed manner for some
reason. I'm not entirely sure why that is. It might have been intended
as a failsafe in case the player managed to get two clues at once,
but you're only able to get two clues at once *because* of the delayed
update.