Plugged up several "dist < bdist" holes, fixed ai.hyperspace() docs.
The ai.hyperspace() docs incorrectly said that it returned a distance
if it failed, when it actually returns one of three error codes, so
I fixed that.

Also, while looking into an occasional problem where ships would
get stuck forever rotating in a circle while trying to hyperspace,
I noticed a particular hole: it only proceeded to the "braking" step
if the distance to the jump was less than a calculated braking distance.
Well, here's a major problem with that: if the ship is already *exactly*
at that point, that distance is zero, while if the ship is already
*exactly* fully stopped, that braking distance is zero. That condition
(the distance and braking distance) were not accounted for at all.
This very well may be *the* source of those problems I've noticed.

I saw several other cases of code only proceeding to the braking step
if the distance to the target was *less than* the braking distance
(impossible if fully stopped), so I patched those up as well, changing
them to less than *or equal to* checks. I can't 100% verify this, but
this hopefully might completely patch up those weird AI errors.