Changed the escorts event to use the "escort" AI, removed "escort_player".
The "escort_player" AI, confusingly, is not used for the player's
escorts; the "escort" AI is, instead. It seems something very weird
happened just over 10 years ago that led to escort.lua being created
and supplanting escort_player.lua, during a commit that purported to
just be moving the Lua code from one place to another.