Auto-aiming now disengages when the player's manual aiming is better.
This is a different alternative to Naev's "disable tracking" option:
rather than having the player decide on their own, this automatically
chooses whichever is more accurate: the auto-aiming available, or
manual aiming. Based on my tests with a Kestrel, this seems to work
very well.

Incidentally, this also removes a redundant line which seems to have
merely been left over from a prior implementation (and didn't do
anything as of 708fb09).