Adjusted the ASAW calculation to be more clear and better commented.
The previous version worked just fine, but there were a couple sort
of kludgy things it did which made it much more difficult to understand
than it needed to be. Part of that was that I was writing days as their
calendar day, and not their zero-indexed day, and there was stuff that
made that work; and another part was just that comments were a bit
unclear in general.

The actual effect of the code here is identical, but it just doesn't
look like some kind of magical wizardry anymore.