Added support for quitting the game while loading is in progress.
This, essentially, cuts down the loading sequence to only what is
absolutely necessary to enter into the game and then immediately quit,
if a quit is requested by the user (e.g. Alt+F4 or the window close
button). This avoids the annoyance of having to wait for all of the
loading to complete before quitting the game if, for example, you
accidentally launch it.

The quit-while-loading process isn't as clean as quitting while the
game is running, but I've done it such that it at least doesn't
segfault and doesn't introduce memory leaks. I've also done my best
to minimize warnings printed when this is done.

One other nice side effect of this change is that the OS gets polled
during loading much more often than previously, which should reduce
the chance of the OS thinking that the game is unresponsive.