Removed some specific requirements, made Shadow require Mercenary License.
This removes the specific requirement to do the Space Family mission
before the Teenager mission, and removes the specific requirement to
do the Teenager mission before the Space Family mission. I'm figuring
that since the Space Family mission is guaranteed to show up at least
once, that can serve as the basic boarding tutorial, and the Shadow
campaign doesn't require disabling ships at the moment.