Removed the last two Coming Out missions and the Coming Out static factions.
The last two missions in the Coming Out campaign are filler that I don't
think is needed. In the reworked Coming Out campaign, the next mission
will be an official military contract, and the next time the player
sees Chelsea will be after they get themself a decked-out combat ship.

While I was doing this, I also got rid of another faction called
"Raiders" which was used by an old, deleted faction, and added a
sentence to flf_dvk02 about the fact that Ion Cannons can be bought
at Sindbad.