Added code which causes non-aggressive AI and player escorts to local jump.
This makes AI consider using a local jump as a last-ditch effort to
survive if armor reaches a certain threshold. It's been defined for
civilians, independents, miners, traders, and the player's escorts.

It is intentionally *not* being defined for factional ships or pirates,
since these are the ships that a player is most likely to target
without trying to board them. We don't want to annoy the player with
this.

The purpose of the AI's use of the local jump is twofold: firstly,
it can help the player's escorts avoid getting killed. Secondly, by
making the ships the player likely doesn't want to kill (even if they
happen to be a pirate) use the escape jump to save their own butts,
that communicates to the player that this is an option for them and
a use of the local jump feature.