Changed the default keybinds, fixed Aristodrink second OSD entry.
I changed the keybinds so that Q is now Escape Jump and E is now
autoface. It seems to not have the same problem that setting E to
Escape Jump had; I actually entirely forgot about what I was testing
and ended up doing another fix at the same time.

That other fix is just a correction to the Aristodrink second OSD
entry.