Fixed a memory leak and added a warning for an error case.
The warning functionally doesn't change anything, but will show us
if there is ever a problem with the static choice_result variable
(which is supposed to be set to NULL before its value is passed on
to the caller of dialogue_runChoice).

The memory leak was in the Options window, when setting keys to the
defaults. The return value of dialogue_runChoice wasn't being freed
like it was supposed to be.