Changed the Prince mission's hostiles spawning, reworded a missed "Empire worker"
I've changed the Baron Prince mission so that it claims all of the
systems it uses and then only spawns artifact hunters in those systems.
This prevents a possible edge case where the behavior could be done
in a system claimed by another mission, potentially causing problems.

Also replaced an instance of "Empire worker" with "Imperial worker"
which I missed previously.