Removed a conditional that caused planet music to suddenly interrupt.
I think we put this in in an effort to fix the bug that was fixed by
c56197a. It wasn't the correct way to fix it and with that in place
I'm pretty sure that particular music bug won't show up again. Fading
out the music rather than suddenly stopping it sounds much better,
so let's switch it back.