Don't allow local jump from AI pilots if it would leave them unable to jump.
This was something I was originally going to do but decided to try it
without this check. It pretty much works fine, but there's an edge-case
where this could make the trader escort mission agonizingly slow, and
plus the fact that fighter bay fighters wouldn't know what to do.
The minor benefit of one extra excape jump for AI pilots probably isn't
worth the downside of possibly running into these edge cases.