Made the close button still ask confirmation in Options and load menus.
This entailed actually implementing use of MENU_OPTIONS (which it
turns out wasn't in use at all), which entailed making the menu_Open
and menu_Close functions public. Also added a new menu flag (MENU_LOAD)
and enabled/disabled it with the Load menu.

Now, while it will insta-quit if you press the window's close button
in the main menu, that doesn't count if you have the Options menu
or Load Game menu opened. This matters for the Options menu since
quitting while it's open would implicitly save the changes you make.
It matters less for the Load Game menu.