Made the extra-careful check for enemy presence slightly less careful.
There's two parts of the change: first, the basic distance enemy weapon
ranges are checked against is the full distance, and the more careful
threshold is 90% of the distance, rather than 90% and 80% respectively.
Second, the more cautious comparison is only used if the pilot is both
actively hostile *and* can see the pilot, rather than if either is
true. This should reduce unnecessary slowdowns without sacrificing
safety.