Changed a few cases of manual table-appendings to table.insert().
This isn't terribly important, but for a long time we didn't know
about table.insert() and did the same function with things like
`T[#t + 1] = x`. Luckily, we spaced it pretty consistently, so it's
easy to find. Best to switch these to table.insert() in the long-term,
though there's no huge rush there.