Made comingout4 more likely to appear, avoid duplicate claims.
Avoiding duplicate claims doesn't seem to actually be necessary based
on how the code looks, but I'm not a fan of allowing it to list the
same systems for claiming twice, aesthetically if nothing else.

This also adds a new "tablehelper" library, which includes the
unused enumerate function and an "inTable" function.