Added medium and large Za'lek test engines.
Seemingly a simple task, but because of the way the Za'lek test
engines are used, it involved changing a lot of things.

Instead of one "Za'lek Test Engine", there are now three separate
engines:

* Za'lek S300 Test Engine
* Za'lek M1200 Test Engine
* Za'lek L6500 Test Engine

For now, they are all similar in comparison to their similar-sized
peers. We might differentiate them more at some point.

Test engine event was modified to work with all three test engines.

The Za'lek Test mission had to be reworked. I decided to make it so
that it explains to you up-front that an engine type is selected
based on the size of your ship, and from there your deposit is
determined. To do that, I had to look into how to check the ship's
slots, and I found the API to do so lacking, so I improved that on
the way.

Thinking about it now, I'm not 100% sure about this method for
the Za'lek Test mission (particularly because if you're flying a
large ship you have to pay a hefty deposit), so maybe I'll adjust it
later. It'll do for now, though.