Sizable overhaul which adds functionality to cut out playing sounds.
Turns out the work needed to finish #4 wasn't all that much. All I
had to do was add an extra attribute to voice structs, and the effort
was all good to go.

That said, positional sounds required some further changes. I never
noticed this before today, but the previous system for some reason
set positioning and rotation as if you were inside your ship… and
lying face-down on the floor of the ship. This meant there was a
rather obnoxious effect where many sounds from your own ship would
play as if they were coming directly from your left or right, which
wasn't too noticeable when infinite sounds were allowed to play, but
once I implemented cut-offs of previously playing instances, it was
pretty bad. The solution to this was to move the listener in the Z
direction toward the player, and while I was at it, I also got rid
of that incredibly silly direction setting. Naikari doesn't rotate
its world, so trying to make the sound direction mimic the player's
ship is not only not needed, it's actively confusing.

There is only one problem that I witnessed in testing, and that is
simply that the beam sounds are a little quiet (likely because their
volume was set too low on purpose so that they wouldn't constantly
sound obnoxiously loud). Will look into a fix for that later.

Closes #4