Added proper definitions for length of sound and voice ID types.
Was previously using the int type, which is only guaranteed to be
at least 16-bit. That's not really a problem for the sound IDs, but
for voice IDs, that could easily be a problem, as a 16-bit integer
has only has a range in the tens of thousands. Making it always
32-bit makes the system a bit more robust.