Replaced the in-house vasprintf with a copy of SDL's.
It looks like they discovered in Naev that the in-house vasprintf
is broken. bobbens' response to that was to use SDL_vasprintf and
require SDL 2.0.18, but the latest stable Debian has only version
2.0.14, so instead, I've decided to just copy Cameron Cawley's work
into our own source code. In doing so, I slightly modified it to use
standard C functions (malloc, realloc, free) rather than SDL's
in-house versions, and got rid of the explicit cast in the malloc
call (which is only in SDL for the benefit of C++ compatibility,
something we don't need to worry about).