Like Naev, removed the custom nsetenv and just use SDL_setenv().
Since bobbens didn't explain why this is OK, I'll do so: the reason
nsetenv was re-added in the past (and why bobbens added the warning
about SDL_setenv) was because they had attempted to use SDL_setenv
for the LANGUAGE environment variable. That is no longer something
that's done, and in fact, in Naikari, the only nsetenv call left is
specific to POSIX systems (e.g. Linux).