Remove checks for setenv/putenv, simplify strndup check.
The environment variable functions aren't used anymore, so we don't
need to check for them.

There's an extra careful strndup check referencing an issue that was
supposedly fixed in Meson 0.54.2. We require at least Meson 0.55, so
that shouldn't be a problem. Will know for sure when I see the CI
results from this commit.