Replaced an incredibly outdated, disabled MELEMENT check.
Looks like MELEMENT checks were created at least 16 years ago, and
entirely disabled at that time because it "shouldn't be needed atm".
No explanation in the Git commit that added this change, but I'm
guessing it's because commodities used to include junk commodities
for missions. Those are gone now, so implemented a new series of
checks for missing elements of commodities. Decided that the required
elements are "name", "description", "price", and "gfx_store".