Imported Naev's current cust_pcall in nlua_tk.c, erased commented-out code.
Naev's current version of cust_pcall seems to be fine, so going with
that. (It actually won't affect us since we don't use Lua custs, but
who knows, maybe we eventually will.)

Also found a case where bobbens replaced some code, but decided to
comment it out with #if 0 rather than just erasing it for some reason.
I've gone ahead and erased that commented-out code.