Added an option to force fallback for LuaJIT.
While we couldn't 100% confirm this, we believe the random crashes
in the release build are due to an old version of LuaJIT found in
the Debian 10 repository. We considered checking if we could just
not install LuaJIT, but a more elegant solution would be to force
it to fallback to the subproject on this end.