Overhauled the way button keys are done, copying wxWidgets and similar.
The Naev way of specifying a key in C and then automatically marking
the first matching letter does work, but it misses out on a bit of
flexibility that manual indication of the proper letter allows. Of
particular note, being able to ensure that the most reasonable letter
is highlighted (as in "Steal All", where the "A" in "Steal" was getting
highlighted despite the "A" in "All" being a better choice), and
enabling translators to change what the hotkey is so that it makes
more sense for the target language (rather than hotkeys being locked
to something that makes sense for English).

The way I implemented it allows any button to end up having a hotkey
if one is defined with the new & notation, and the hotkey can change
when the button text is changed with window_buttonCaption. I could
have decided not to do this, but I figured allowing this sort of thing
is more likely to be a benefit than a detriment. The only real downside
to this design is that including the & character in buttons is less
permissive (but still possible as long as it is followed by a space or
a UTF-8 code point which isn't ASCII).