Got rid of some dangerous hacks, made debug builds require _GNU_SOURCE.
There were two stages to this: first, debug_enableFPUExcept() was
moved to a separate file to enable isolating use of _GNU_SOURCE.
Then, because this isolated use of _GNU_SOURCE would lead to use of
GNU extensions that require it not working, when another dependency
on _GNU_SOURCE was added by 2e9ca71d5c7caf2a12dd7e3526a67d65caaf4893,
__USE_GNU was defined instead as a hack since _GNU_SOURCE didn't work
anymore (the reason being that the way GNU C makes use of _GNU_SOURCE
is to include a header which uses its value to define __USE_GNU,
among other things). This behavior is potentially very dangerous and
not in-line with how GNU extensions are supposed to be used.

To replace this, I've made it so that for debug builds, _GNU_SOURCE
gets defined by the top-level meson.build instead, and documented
why _GNU_SOURCE is needed (to the best of my knowledge).