Fixed a bug where player inputs could be swallowed by tk.
This is a longstanding issue, but was kind of hidden/mitigated by
the fact that tk would pass on inputs that it wasn't using (that is,
keys which weren't used as hotkeys) to the main input handler. While
that behavior certainly helped, it would mean that if, for example,
the acceleration key *is* a hotkey, it does get swallowed and if the
window opened while it was pressed, the player would see constant
acceleration until they re-pressed and released the acceleration key.
The mouse controls were also unaffected by this mitigation, since
whenever it was added the mouse buttons weren't used to control thrust.

To ensure there's never a problem again, all tk windows now clear all
player input when they open, behaving as if the player released all
keys. This means if you want to hold a key to continue acceleration
after the message, you first need to re-press it after the window is
opened, which is a small price to pay for making sure no unexpected
acceleration happens.