Fixed broken error checking for conditionals in missions/events.
It was expecting some specific error codes, and the result was that
this error checking actually caused all errors (including syntax
errors, like one I experienced moments ago) to be treated as a
success of the condition, which was the opposite of what we want in
a scenario like that, while failing to print a warning.

This also enhances the error by including a debug printout of the
actual text where the error occurred.