Changed our subtitle to "Eye of Chaos".
Deliberated over this for quite a bit, but I think of all the
possibilities we could think of, I like "Eye of Chaos" the best.

At some point, the plan is to write some characters which describe
the then-current situation using the subtitle. For example:

* "We're in the eye of looming chaos."
* "The Empire is in the eye of chaos and they don't even know it."
* "This stability is a farce. It's just the eye of impending chaos."

Also included in this commit a change that doesn't really matter,
but I saw that some (never actually executed) code would for some
reason recurse on the creation of a new character for the player if
the player didn't enter a ship name where required. That just seems
kind of silly to me. If they clicked the "Cancel" button, they
probably changed their mind on creating the new game.