Improved save on load timing and got rid of refusing to save on places.
This causes it to not save the game until you actually load a game,
which avoids repeated saves that cause backups to be needlessly
overridden. I also got rid of the old conditional for all save_all()
calls that first checked to make sure the planet you were landed on
had a refuel service, as this protection feature is now obsolete.
(Actually, I had thought that we already removed it, but I guess it
had redundant checks we failed to notice.)