Revert "Use pilot's total value rather than ship price for credits and bounty."
This reverts commit 8e5ae8ea3710cef07dca592b88557f18e058189f.

I've decided against it. Switching to using value as the heuristic,
while in theory sounding more accurate, in practice isn't going to
make much difference especially compared to the massive work of
recalibrating everything and retesting. At the end of the day,
whether it's ship base price or total ship + outfits value, it's
still just a heuristic.