Nerfed Photon Dagger's damage (making it the same as the Banshee Rocket).
We've been using it a lot in recent playtesting, and it looks like
we massively underestimated how big of a benefit its high speed is.
Having double the damage over the Banshee makes the Banshee look
useless by comparison, which wasn't what we were going for with the
Photon Dagger.

In fact, the Photon Dagger is so incredibly good that we might want
to lower its damage even moreso.