Further improved handling of instant mode change and fixed a memory error.
We made a bit of a mistake: pilot_weapSetAdd depends on the caller
to ensure the level set is valid and not -1. Failing to include this
check caused a memory access error.

Also made it so that it not only checks that a change weapset exists,
but that it has weapons on it, before allowing a change of the
weapon set to instant mode. This should prevent confusing situations
where an empty switch weapset exist and gets automatically switched
to.