Added hilighting of missions in the selected system.
This commit also includes a fix for a basic mistake we made in the
last commit of forgetting to check that we're actually landed before
performing the land updates.

As a matter of necessity, this commit also removes the long-standing
behavior of moving the map based on mission selection, since that
runs counter to the idea of allowing the player to choose missions
by location.

Next on the list is to make all potential missions' mission markers
show at once, rather than only showing the markers of the selected
mission.